requirements.txt
firebase-admin
Flask
pydantic
pytest
Pillow


